refactor(signinstorelogic.ts): decompose common code signUp & signIn in signInStoreLogic.ts
