+++ src/pages/App/tests/commercetools/middlewareOptions.test.ts
@@ -3,9 +3,7 @@ import {
   httpMiddlewareOptions,
   passwordAuthMiddlewareOptions,
 } from 'shared/libs/commercetools/middlewareOptions';
-import {
-  describe, test, assert,
-} from 'vitest';
+import { describe, test, assert } from 'vitest';
 
 describe('Middleware options tests', () => {
   test('httpMiddlewareOptions should have correct host', () => {
@@ -13,12 +11,24 @@ describe('Middleware options tests', () => {
   });
 
   test('passwordAuthMiddlewareOptions should have correct host and projectKey', () => {
-    assert.equal(passwordAuthMiddlewareOptions.host, import.meta.env.VITE_CTP_AUTH_URL);
-    assert.equal(passwordAuthMiddlewareOptions.projectKey, import.meta.env.VITE_CTP_PROJECT_KEY);
+    assert.equal(
+      passwordAuthMiddlewareOptions.host,
+      import.meta.env.VITE_CTP_AUTH_URL,
+    );
+    assert.equal(
+      passwordAuthMiddlewareOptions.projectKey,
+      import.meta.env.VITE_CTP_PROJECT_KEY,
+    );
   });
 
   test('anonymousAuthMiddlewareOptions should have correct host and projectKey', () => {
-    assert.equal(anonymousAuthMiddlewareOptions.host, import.meta.env.VITE_CTP_AUTH_URL);
-    assert.equal(anonymousAuthMiddlewareOptions.projectKey, import.meta.env.VITE_CTP_PROJECT_KEY);
+    assert.equal(
+      anonymousAuthMiddlewareOptions.host,
+      import.meta.env.VITE_CTP_AUTH_URL,
+    );
+    assert.equal(
+      anonymousAuthMiddlewareOptions.projectKey,
+      import.meta.env.VITE_CTP_PROJECT_KEY,
+    );
   });
 });
